login with user, add pet for user, authentication server actions

+++ src/actions/actions.ts
@@ -1,14 +1,48 @@
 "use server";
 
+import { signIn, signOut } from "@/lib/auth";
 import prisma from "@/lib/db";
-import { PetEssentials } from "@/lib/types";
 import { sleep } from "@/lib/utils";
 import { petFormSchema, petIdSchema } from "@/lib/validations";
-import { Pet } from "@prisma/client";
 import { revalidatePath } from "next/cache";
+import bcrypt from "bcryptjs";
+import { redirect } from "next/navigation";
+import { checkAuth } from "@/lib/server-utils";
 
+//--------- user actions ------------
+export async function logIn(formData: FormData) {
+  // const authData = Object.fromEntries(formData.entries());
+
+  await signIn("credentials", formData);
+
+  redirect("/app/dashboard");
+}
+
+export async function signUp(formData: FormData) {
+  // const authData = Object.fromEntries(formData.entries());
+
+  const hashedPassword = await bcrypt.hash(
+    formData.get("password") as string,
+    10
+  );
+  await prisma.user.create({
+    data: {
+      email: formData.get("email") as string,
+      hashedPassword,
+    },
+  });
+
+  await signIn("credentials", formData);
+}
+
+export async function logOut() {
+  await signOut({ redirectTo: "/" });
+}
+
+// --------pet actions ----------
 export async function addPet(pet: unknown) {
   await sleep(1000);
+  const session = await checkAuth();
   const validatedPet = petFormSchema.safeParse(pet);
 
   if (!validatedPet.success) {
@@ -18,7 +52,14 @@ export async function addPet(pet: unknown) {
   }
   try {
     await prisma.pet.create({
-      data: validatedPet.data,
+      data: {
+        ...validatedPet.data,
+        user: {
+          connect: {
+            id: session.user.id,
+          },
+        },
+      },
     });
   } catch (err) {
     return {
@@ -32,6 +73,9 @@ export async function addPet(pet: unknown) {
 export async function editPet(petId: unknown, newPet: unknown) {
   await sleep(1000);
 
+  //authentication
+  const session = await checkAuth();
+
   const validPetId = petIdSchema.safeParse(petId);
   const validPetData = petFormSchema.safeParse(newPet);
 
@@ -41,6 +85,25 @@ export async function editPet(petId: unknown, newPet: unknown) {
     };
   }
 
+  //authorization
+  const pet = await prisma.pet.findUnique({
+    where: {
+      id: validPetId.data,
+    },
+  });
+
+  if (!pet) {
+    return {
+      message: "No pet found",
+    };
+  }
+
+  if (pet.userId !== session.user.id) {
+    return {
+      message: "Not authorized",
+    };
+  }
+  //database mutation
   try {
     await prisma.pet.update({
       where: {
@@ -60,6 +123,10 @@ export async function editPet(petId: unknown, newPet: unknown) {
 export async function deletePet(petId: unknown) {
   await sleep(1000);
 
+  //authentication
+  const session = await checkAuth();
+
+  //validation
   const validPetId = petIdSchema.safeParse(petId);
 
   if (!validPetId.success) {
@@ -68,6 +135,26 @@ export async function deletePet(petId: unknown) {
     };
   }
 
+  //authorization
+  const pet = await prisma.pet.findUnique({
+    where: {
+      id: validPetId.data,
+    },
+  });
+
+  if (!pet) {
+    return {
+      message: "No pet found",
+    };
+  }
+
+  if (pet.userId !== session.user.id) {
+    return {
+      message: "Not authorized",
+    };
+  }
+
+  //database mutaion
   try {
     await prisma.pet.delete({
       where: {
